requirements.txt
Django==4.2.2
Pillow==9.5.0
psycopg2==2.9.6
djangorestframework==3.14.0